+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `shouldForceWrapper` and `shouldForceBlock` to `Markdown`
 - [PATCH] Corrections for how height is calculated in `MultiLineInput`
 - [MAJOR] Removed `Head` in favor of react 19 inline metadata tags
+- [MINOR] Added `isResizableVertically` and `isResizableHorizontally` to `MultiLineInput`
 
 ### Removed